Moved GetOptionsWithHeaders to MediaFileUrlHelper static class
