+++ MediaManager.iOS/MediaFileUrlHelper.cs
@@ -1,4 +1,8 @@
-﻿using Foundation;
+﻿using System.Collections.Generic;
+
+using Foundation;
+using AVFoundation;
+
 using Plugin.MediaManager.Abstractions;
 using Plugin.MediaManager.Abstractions.Enums;
 
@@ -13,5 +17,24 @@ namespace Plugin.MediaManager
 
             return url;
         }
+
+		public static AVUrlAssetOptions GetOptionsWithHeaders(IDictionary<string, string> headers)
+		{
+			var nativeHeaders = new NSMutableDictionary();
+
+			foreach (var header in headers)
+			{
+				nativeHeaders.Add((NSString)header.Key, (NSString)header.Value);
+			}
+
+			var nativeHeadersKey = (NSString)"AVURLAssetHTTPHeaderFieldsKey";
+
+			var options = new AVUrlAssetOptions(NSDictionary.FromObjectAndKey(
+				nativeHeaders,
+				nativeHeadersKey
+			));
+
+			return options;
+		}
     }
 }
